portfolio_backend/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
celery
redis
requests
yfinance
pandas
numpy
scikit-learn
joblib
transformers
torch
feedparser
textblob
vaderSentiment
finnhub-python
praw
newsapi-python
python-dotenv
